+++ portfolio/src/main/webapp/pages/server-dev.html
@@ -49,8 +49,8 @@
       <main>
         <section class="content">
           <h2>Add a new comment!
-          <form action="/comment-data" method="POST">
-            <input type="text" name="new-comment"><br/>
+          <form action="/new-comment" method="POST">
+            <input type="text" name="comment"><br/>
             <input type="submit" /><br/>
           </form>
         </section>
